Move tests in test_views
